タイマー画面のコンポーネント化
