+++ lib/scenes/timer/timer_screen.dart
@@ -1,0 +1,48 @@
+import 'dart:async';
+import 'package:flutter/material.dart';
+
+class TimerScene extends StatefulWidget {
+  const TimerScene({super.key});
+
+  @override
+  _TimerScreenState createState() => _TimerScreenState();
+}
+
+class _TimerScreenState extends State<TimerScene> {
+  int totalSeconds = 0;
+  late Timer _timer;
+
+  @override
+  void initState() {
+    super.initState();
+    startTimer();
+  }
+
+  @override
+  void dispose() {
+    _timer.cancel();
+    super.dispose();
+  }
+
+  void startTimer() {
+    _timer = Timer.periodic(const Duration(seconds: 1), (timer) {
+      setState(() {
+        totalSeconds++;
+      });
+    });
+  }
+
+  String formatTime(int totalSeconds) {
+    int minutes = totalSeconds ~/ 60;
+    int seconds = totalSeconds % 60;
+    return '${minutes.toString().padLeft(2, '0')}:${seconds.toString().padLeft(2, '0')}';
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Text(
+      formatTime(totalSeconds),
+      style: const TextStyle(fontSize: 28, fontWeight: FontWeight.bold),
+    );
+  }
+}
